PDFRegionDecoder.init now returns the maximum page width + sum of page heights
Previously it returned the width and height of the first page. This worked fine for PDF documents where all pages had the same size. But for documents with different page size it lead to incorrect results. The more precise way takes a bit longer to calculate, but the "costs" seem acceptable even for PDF documents with multiple hundred pages.